Created models package, changed Server package to lower case
