+++ Server/Server/api/__init__.py
@@ -4,4 +4,4 @@ from flask import Blueprint
 api_blueprint = Blueprint('api', __name__, url_prefix='/api')
 api = Api(api_blueprint)
 
-from Server.api import auth, clasrooms, reports, student_status
+from server.api import auth, clasrooms, reports, student_status
