+++ Server/run.py
@@ -1,8 +1,8 @@
-from Server import app, db
+from server import app, db
 import os
 
 if __name__ == '__main__':
-    if not os.path.exists("/Server/site.db"):
+    if not os.path.exists("/server/site.db"):
         db.create_all()
         db.session.commit()
     app.run(debug=True)
